update-12-11-2020

+++ src/main/java/fr/inra/sad/bagap/chloe/counting/QuantitativeCounting.java
@@ -14,7 +14,7 @@ public class QuantitativeCounting extends Counting implements QuantitativeCounti
 		this.theoriticalSize = theoriticalSize;
 	}
 	
-	public void setCounts(double[] counts){
+	public void setCounts(float[] counts){
 		totalValues = 0;
 		validValues = 0;
 		sum = 0;
